Prompt user that a TK window needs attention
On windows (at least windows 7 computers I have worked with), the TK window
will not raise to the front with `tkfocus`.  This can lead to hard-to-diagnose
issues where a password is being prompted for but the user does not know this.
Other platforms may have similar issues, as I believe that window focussing
is not compulsary and the window will not be treated like a modal dialog (vs
the rstudioapi approach which is really slick).

This commit just adds a small message to the console indicating that the TK
window will need some input.

(there is also a fix of a partial `$` match of OS -> OS.type)